Crate polars_arrow

Modules

array

Contains the Array and MutableArray trait objects declaring arrays, as well as concrete arrays (such as Utf8Array and MutableUtf8Array).

bitmap

contains Bitmap and MutableBitmap, containers of bool.

buffer

Contains Buffer, an immutable container for all Arrow physical types (e.g. i32, f64).

compute

contains a wide range of compute operations (e.g. [arithmetics], aggregate, [filter], [comparison], and [sort])

datatypes

Contains all metadata, such as PhysicalType, ArrowDataType, Field and ArrowSchema.

ffi

contains FFI bindings to import and export Array via Arrow’s C Data Interface

io

legacy

mmap

Memory maps regions defined on the IPC format into Array.

offset

Contains the declaration of Offset

pushable

record_batch

Contains RecordBatchT, a container of Array where every array has the same length.

scalar

contains the Scalar trait object representing individual items of Arrays, as well as concrete implementations such as BooleanScalar.

storage

temporal_conversions

Conversion methods for dates and times.

trusted_len

Declares TrustedLen.

types

Sealed traits and implementations to handle all physical types used in this crate.

util

Misc utilities used in different places in the crate.

Macros

match_integer_type

with_match_primitive_type

with_match_primitive_type_full

Enums

Either

The enum Either with variants Left and Right is a general purpose sum type with two cases.