Struct polars_arrow::array::MutableBinaryViewArray

pub struct MutableBinaryViewArray<T: ViewType + ?Sized> { /* private fields */ }

Implementations

impl<T: ViewType + ?Sized> MutableBinaryViewArray<T>

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn views_mut(&mut self) -> &mut Vec<View>

pub fn views(&self) -> &[View]

pub fn completed_buffers(&self) -> &[Buffer<u8>]

pub fn validity(&mut self) -> Option<&mut MutableBitmap>

pub fn reserve(&mut self, additional: usize)

Reserves additional elements and additional_buffer on the buffer.

pub fn len(&self) -> usize

pub fn capacity(&self) -> usize

pub unsafe fn push_view(&mut self, v: View, buffers: &[Buffer<u8>])

Safety

caller must allocate enough capacity

caller must ensure the view and buffers match.

pub fn push_value_ignore_validity<V: AsRef<T>>(&mut self, value: V)

pub fn push_value<V: AsRef<T>>(&mut self, value: V)

pub fn push<V: AsRef<T>>(&mut self, value: Option<V>)

pub fn push_null(&mut self)

pub fn extend_null(&mut self, additional: usize)

pub fn extend_constant<V: AsRef<T>>( &mut self, additional: usize, value: Option<V>, )

pub fn with_validity(self, validity: Option<MutableBitmap>) -> Self

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<MutableBitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn apply_validity<F: FnOnce(MutableBitmap) -> MutableBitmap>( &mut self, f: F, )

Applies a function f to the validity of this array.

This is an API to leverage clone-on-write

Panics

This function panics if the function f modifies the length of the [Bitmap].

pub fn extend_values<I, P>(&mut self, iterator: I)
where I: Iterator<Item = P>, P: AsRef<T>,

pub fn extend_trusted_len_values<I, P>(&mut self, iterator: I)
where I: TrustedLen<Item = P>, P: AsRef<T>,

pub fn extend<I, P>(&mut self, iterator: I)
where I: Iterator<Item = Option<P>>, P: AsRef<T>,

pub fn extend_trusted_len<I, P>(&mut self, iterator: I)
where I: TrustedLen<Item = Option<P>>, P: AsRef<T>,

pub fn from_iterator<I, P>(iterator: I) -> Self
where I: Iterator<Item = Option<P>>, P: AsRef<T>,

pub fn from_values_iter<I, P>(iterator: I) -> Self
where I: Iterator<Item = P>, P: AsRef<T>,

pub fn from<S: AsRef<T>, P: AsRef<[Option<S>]>>(slice: P) -> Self

pub fn freeze(self) -> BinaryViewArrayGeneric<T>

pub fn value(&self, i: usize) -> &T

pub unsafe fn value_unchecked(&self, i: usize) -> &T

Returns the element at index i

Safety

Assumes that the i < self.len.

pub fn values_iter(&self) -> ArrayValuesIter<'_, MutableBinaryViewArray<T>>

Returns an iterator of &[u8] over every element of this array, ignoring the validity

impl MutableBinaryViewArray<[u8]>

pub fn validate_utf8( &mut self, buffer_offset: usize, views_offset: usize, ) -> PolarsResult<()>