Struct polars_arrow::array::MutableDictionaryArray

pub struct MutableDictionaryArray<K: DictionaryKey, M: MutableArray> { /* private fields */ }

Implementations

impl<K: DictionaryKey, M: MutableArray + Default> MutableDictionaryArray<K, M>

pub fn new() -> Self

Creates an empty MutableDictionaryArray.

impl<K: DictionaryKey, M: MutableArray> MutableDictionaryArray<K, M>

pub fn try_empty(values: M) -> PolarsResult<Self>

Creates an empty MutableDictionaryArray from a given empty values array.

Errors

Errors if the array is non-empty.

pub fn from_values(values: M) -> PolarsResult<Self>where M: Indexable, M::Type: Eq + Hash,

Creates an empty MutableDictionaryArray preloaded with a given dictionary of values. Indices associated with those values are automatically assigned based on the order of the values.

Errors

Errors if there’s more values than the maximum value of K or if values are not unique.

pub fn into_empty(self) -> Self

Creates an empty MutableDictionaryArray retaining the same dictionary as the current mutable dictionary array, but with no data. This may come useful when serializing the array into multiple chunks, where there’s a requirement that the dictionary is the same. No copying is performed, the value map is moved over to the new array.

pub fn to_empty(&self) -> Selfwhere M: Clone,

Same as into_empty but clones the inner value map instead of taking full ownership.

pub fn push_null(&mut self)

pushes a null value

pub fn values(&self) -> &M

returns a reference to the inner values.

pub fn into_arc(self) -> Arc<dyn Array>

converts itself into Arc<dyn Array>

pub fn into_box(self) -> Box<dyn Array>

converts itself into Box<dyn Array>

pub fn reserve(&mut self, additional: usize)

Reserves additional slots.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the MutableDictionaryArray to fit its current length.

pub fn keys(&self) -> &MutablePrimitiveArray<K>

Returns the dictionary keys