polars_arrow::array

Trait PushUnchecked

pub trait PushUnchecked<A> {
    // Required method
    unsafe fn push_unchecked(&mut self, item: A);
}

A trait describing the ability of a struct to receive new items.

Required Methods

unsafe fn push_unchecked(&mut self, item: A)

Push a new element that holds the invariants of the struct.

Safety

The items must uphold the invariants of the struct Read the specific implementation of the trait to understand what these are.