polars_arrow::array

Trait TryExtend

pub trait TryExtend<A> {
    // Required method
    fn try_extend<I: IntoIterator<Item = A>>(
        &mut self,
        iter: I,
    ) -> PolarsResult<()>;
}

A trait describing the ability of a struct to create itself from a iterator. This is similar to Extend, but accepted the creation to error.

Required Methods

fn try_extend<I: IntoIterator<Item = A>>(&mut self, iter: I) -> PolarsResult<()>

Fallible version of Extend::extend.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.