polars_arrow

Module bitmap

contains Bitmap and MutableBitmap, containers of bool.

Re-exports

pub use iterator::IntoIter;

Modules

aligned

bitmask

iterator

utils

General utilities for bitmaps representing items where LSB is the first item.

Structs

Bitmap

An immutable container semantically equivalent to Arc<Vec<bool>> but represented as Arc<Vec<u8>> where each boolean is represented as a single bit.

BitmapBuilder

Used to build bitmaps bool-by-bool in sequential order.

MutableBitmap

A container of booleans. MutableBitmap is semantically equivalent to Vec<bool>.

Enums

OptBitmapBuilder

A wrapper for BitmapBuilder that does not allocate until the first false is pushed. Less efficient if you know there are false values because it must check if it has allocated for each push.

Functions

and

Compute bitwise A AND B operation.

and_not

Compute bitwise A AND NOT B operation.

binary

Apply a bitwise operation op to two inputs and return the result as a Bitmap.

binary_assign

Apply a bitwise binary operation to a MutableBitmap.

binary_assign_mut

Apply a bitwise binary operation to a MutableBitmap.

binary_fold

Apply a bitwise operation op to two inputs and fold the result.

binary_fold_mut

Apply a bitwise operation op to two inputs and fold the result.

chunk_iter_to_vec

Creates a Vec<u8> from a TrustedLen of BitChunk.

intersects_with

intersects_with_mut

num_edges

num_intersections_with

or

Compute bitwise A OR B operation.

or_not

Compute bitwise A OR NOT B operation.

quaternary

Apply a bitwise operation op to four inputs and return the result as a Bitmap.

select

Compute out[i] = if selector[i] { truthy[i] } else { falsy[i] }.

select_constant

Compute out[i] = if selector[i] { truthy[i] } else { falsy }.

ternary

Apply a bitwise operation op to three inputs and return the result as a Bitmap.

unary

Apply a bitwise operation op to one input and return the result as a Bitmap.

unary_assign

Applies a function to every bit of this MutableBitmap in chunks

xor

Compute bitwise XOR operation.