Struct polars_arrow::bitmap::bitmask::BitMask

pub struct BitMask<'a> { /* private fields */ }

Implementations

impl<'a> BitMask<'a>

pub fn from_bitmap(bitmap: &'a Bitmap) -> Self

pub fn len(&self) -> usize

pub fn split_at(&self, idx: usize) -> (Self, Self)

pub unsafe fn split_at_unchecked(&self, idx: usize) -> (Self, Self)

Safety

The index must be in-bounds.

pub fn get_u32(&self, idx: usize) -> u32

pub fn nth_set_bit_idx(&self, n: usize, start: usize) -> Option<usize>

Computes the index of the nth set bit after start.

Both are zero-indexed, so nth_set_bit_idx(0, 0) finds the index of the first bit set (which can be 0 as well). The returned index is absolute, not relative to start.

pub fn nth_set_bit_idx_rev(&self, n: usize, end: usize) -> Option<usize>

Computes the index of the nth set bit before end, counting backwards.

Both are zero-indexed, so nth_set_bit_idx_rev(0, len) finds the index of the last bit set (which can be 0 as well). The returned index is absolute (and starts at the beginning), not relative to end.

pub fn get(&self, idx: usize) -> bool