Module polars_arrow::compute::aggregate

Contains different aggregation functions

Re-exports

pub use crate::types::PrimitiveType;

Traits

Sum

Object that can reduce itself to a number. This is used in the context of SIMD to reduce a MD (e.g. [f32; 16]) into a single number (f32).

Functions

can_sum

Whether sum supports data_type

estimated_bytes_size

Returns the total (heap) allocated size of the array in bytes.

sum

Returns the sum of all elements in array as a Scalar of the same physical and logical types as array.

sum_primitive

Returns the sum of values in the array.

sum_slice

Compute the sum of a slice