Module polars_arrow::compute::filter

Contains operators to filter arrays such as filter.

Functions

build_filter

Returns a prepared function optimized to filter multiple arrays. Creating this function requires time, but using it is faster than filter when the same filter needs to be applied to multiple arrays (e.g. a multiple columns).

filter

Filters an Array, returning elements matching the filter (i.e. where the values are true).

filter_chunk

Returns a new Chunk with arrays containing only values matching the filter. This is a convenience function: filter multiple columns is embarrassingly parallel.

Type Aliases

Filter

Function that can filter arbitrary arrays