polars_arrow::compute

Module temporal

Defines temporal kernels for time and date related functions.

Functions

day

Extracts the days of a temporal array as PrimitiveArray<i8>.

hour

Extracts the hours of a temporal array as PrimitiveArray<i8>. Value ranges from 0 to 23. Use [can_hour] to check if this operation is supported for the target ArrowDataType.

iso_week

Extracts ISO week of a temporal array as PrimitiveArray<i8>.

minute

Extracts the minutes of a temporal array as PrimitiveArray<i8>. Value ranges from 0 to 59. Use [can_minute] to check if this operation is supported for the target ArrowDataType.

month

Extracts the months of a temporal array as PrimitiveArray<i8>.

nanosecond

Extracts the nanoseconds of a temporal array as PrimitiveArray<i32>.

second

Extracts the seconds of a temporal array as PrimitiveArray<i8>. Value ranges from 0 to 59. Use [can_second] to check if this operation is supported for the target ArrowDataType.

weekday

Extracts weekday of a temporal array as PrimitiveArray<i8>.

year

Extracts the years of a temporal array as PrimitiveArray<i32>.