polars_arrow::compute::temporal

Function minute

pub fn minute(array: &dyn Array) -> PolarsResult<PrimitiveArray<i8>>

Extracts the minutes of a temporal array as PrimitiveArray<i8>. Value ranges from 0 to 59. Use [can_minute] to check if this operation is supported for the target ArrowDataType.