polars_arrow::compute::temporal

Function nanosecond

pub fn nanosecond(array: &dyn Array) -> PolarsResult<PrimitiveArray<i32>>

Extracts the nanoseconds of a temporal array as PrimitiveArray<i32>.

Value ranges from 0 to 1_999_999_999. The range from 1_000_000_000 to 1_999_999_999 represents the leap second. Use [can_nanosecond] to check if this operation is supported for the target ArrowDataType.