polars_arrow

Module datatypes

Contains all metadata, such as PhysicalType, ArrowDataType, Field and ArrowSchema.

Re-exports

pub use crate::types::PrimitiveType;

Modules

reshape

Structs

ExtensionType

Field

Represents Arrow’s metadata of a “column”.

UnionType

Enums

ArrowDataType

The set of supported logical types in this crate.

IntegerType

the set of valid indices types of a dictionary-encoded Array. Each type corresponds to a variant of crate::array::DictionaryArray.

IntervalUnit

Interval units defined in Arrow

PhysicalType

The set of physical types: unique in-memory representations of an Arrow array.

TimeUnit

The time units defined in Arrow.

UnionMode

Mode of ArrowDataType::Union

Statics

DTYPE_CATEGORICAL_LEGACY

DTYPE_CATEGORICAL_NEW

DTYPE_ENUM_VALUES_LEGACY

DTYPE_ENUM_VALUES_NEW

Functions

get_extension

support get extension for metadata

Type Aliases

ArrowSchema

An ordered sequence of Fields

ArrowSchemaRef

IdxArr

Metadata

typedef for BTreeMap<PlSmallStr, PlSmallStr> denoting Field’s and ArrowSchema’s metadata.

SchemaRef

typedef for Arc<ArrowSchema>.