polars_arrow::ffi

Function import_field_from_c

pub unsafe fn import_field_from_c(field: &ArrowSchema) -> PolarsResult<Field>

Imports a Field from the C data interface.

Safety

This function is intrinsically unsafe and relies on a ArrowSchema being valid according to the C data interface (FFI).