polars_arrow::io::ipc

Module write

APIs to write to Arrow’s IPC format.

Structs

DictionaryTracker

Keeps track of dictionaries that have been written, to avoid emitting the same dictionary multiple times. Can optionally error if an update to an existing dictionary is attempted, which isn’t allowed in the FileWriter.

EncodedData

Stores the encoded data, which is an ipc::Schema::Message, and optional Arrow data

FileWriter

Arrow file writer

Record

An array RecordBatchT with optional accompanying IPC fields.

StreamWriter

Arrow stream writer

WriteOptions

Options declaring the behaviour of writing to IPC

Enums

Compression

Compression codec

Functions

commit_encoded_arrays

default_ipc_fields

Assigns every dictionary field a unique ID

dictionaries_to_encode

Find the dictionary that are new and need to be encoded.

encode_array

encode_dictionary

Encode a dictionary array with a certain id.

encode_new_dictionaries

encode_record_batch

Write RecordBatchT into two sets of bytes, one for the header (ipc::Schema::Message) and the other for the batch’s data

schema_to_bytes

Converts a ArrowSchema and IpcFields to a flatbuffers-encoded arrow_format::ipc::Message.

write

Writes an Array to arrow_data