Struct polars_arrow::io::iterator::BufStreamingIterator

pub struct BufStreamingIterator<I, F, T>where
    I: Iterator<Item = T>,
    F: FnMut(T, &mut Vec<u8>),{ /* private fields */ }

A StreamingIterator with an internal buffer of Vec<u8> used to efficiently present items of type T as &[u8]. It is generic over the type T and the transformation F: T -> &[u8].

Implementations

impl<I, F, T> BufStreamingIterator<I, F, T>where I: Iterator<Item = T>, F: FnMut(T, &mut Vec<u8>),

pub fn new(iterator: I, f: F, buffer: Vec<u8>) -> Self