polars_arrow::legacy::array

Trait ListFromIter

pub trait ListFromIter {
    // Provided methods
    unsafe fn from_iter_primitive_trusted_len<T, P, I>(
        iter: I,
        dtype: ArrowDataType,
    ) -> ListArray<i64>
       where T: NativeType,
             P: IntoIterator<Item = Option<T>>,
             I: IntoIterator<Item = Option<P>> { ... }
    unsafe fn from_iter_bool_trusted_len<I, P>(iter: I) -> ListArray<i64>
       where I: IntoIterator<Item = Option<P>>,
             P: IntoIterator<Item = Option<bool>> { ... }
    unsafe fn from_iter_binview_trusted_len<I, P, Ref, T: ViewType + ?Sized>(
        iter: I,
        n_elements: usize,
    ) -> ListArray<i64>
       where I: IntoIterator<Item = Option<P>>,
             P: IntoIterator<Item = Option<Ref>>,
             Ref: AsRef<T> { ... }
    unsafe fn from_iter_utf8_trusted_len<I, P, Ref>(
        iter: I,
        n_elements: usize,
    ) -> ListArray<i64>
       where I: IntoIterator<Item = Option<P>>,
             P: IntoIterator<Item = Option<Ref>>,
             Ref: AsRef<str> { ... }
    unsafe fn from_iter_binary_trusted_len<I, P, Ref>(
        iter: I,
        n_elements: usize,
    ) -> ListArray<i64>
       where I: IntoIterator<Item = Option<P>>,
             P: IntoIterator<Item = Option<Ref>>,
             Ref: AsRef<[u8]> { ... }
}

Provided Methods

unsafe fn from_iter_primitive_trusted_len<T, P, I>( iter: I, dtype: ArrowDataType, ) -> ListArray<i64>
where T: NativeType, P: IntoIterator<Item = Option<T>>, I: IntoIterator<Item = Option<P>>,

Create a list-array from an iterator. Used in group_by agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_bool_trusted_len<I, P>(iter: I) -> ListArray<i64>
where I: IntoIterator<Item = Option<P>>, P: IntoIterator<Item = Option<bool>>,

Create a list-array from an iterator. Used in group_by agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_binview_trusted_len<I, P, Ref, T: ViewType + ?Sized>( iter: I, n_elements: usize, ) -> ListArray<i64>
where I: IntoIterator<Item = Option<P>>, P: IntoIterator<Item = Option<Ref>>, Ref: AsRef<T>,

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_utf8_trusted_len<I, P, Ref>( iter: I, n_elements: usize, ) -> ListArray<i64>
where I: IntoIterator<Item = Option<P>>, P: IntoIterator<Item = Option<Ref>>, Ref: AsRef<str>,

Create a list-array from an iterator. Used in group_by agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_binary_trusted_len<I, P, Ref>( iter: I, n_elements: usize, ) -> ListArray<i64>
where I: IntoIterator<Item = Option<P>>, P: IntoIterator<Item = Option<Ref>>, Ref: AsRef<[u8]>,

Create a list-array from an iterator. Used in group_by agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.