polars_arrow::legacy::kernels

Module set

Functions

scatter_single_non_null

Efficiently sets value at the indices from the iterator to set_value. The new array is initialized with a memcpy from the old values.

set_at_nulls

Set values in a primitive array where the primitive array has null values. this is faster because we don’t have to invert and combine bitmaps

set_with_mask

Set values in a primitive array based on a mask array. This is fast when large chunks of bits are set or unset.