polars_arrow::legacy::kernels

Module take_agg

kernels that combine take and aggregations.

Functions

online_variance

Numerical stable online variance aggregation.

take_agg_bin_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.

take_agg_bin_iter_unchecked_no_null⚠

Take kernel for single chunk and an iterator as index.

take_agg_no_null_primitive_iter_unchecked⚠

Take kernel for single chunk without nulls and an iterator as index.

take_agg_primitive_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.

take_agg_primitive_iter_unchecked_count_nulls⚠

Take kernel for single chunk and an iterator as index.

take_max_bool_iter_unchecked_no_nulls⚠

Take kernel for single chunk and an iterator as index.

take_max_bool_iter_unchecked_nulls⚠

Take kernel for single chunk and an iterator as index.

take_min_bool_iter_unchecked_no_nulls⚠

Take kernel for single chunk and an iterator as index.

take_min_bool_iter_unchecked_nulls⚠

Take kernel for single chunk and an iterator as index.

take_var_no_null_primitive_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.

take_var_nulls_primitive_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.