Type Alias polars_arrow::legacy::prelude::LargeListArray

pub type LargeListArray = ListArray<i64>;

Aliased Type

struct LargeListArray { /* private fields */ }

Implementations

impl<'a, O: Offset> ListArray<O>

pub fn iter( &'a self ) -> ZipValidity<Box<dyn Array>, ListValuesIter<'a, O>, BitmapIter<'a>>

Returns an iterator of Option<Box<dyn Array>>

pub fn values_iter(&'a self) -> ListValuesIter<'a, O>

Returns an iterator of Box<dyn Array>

impl<O: Offset> ListArray<O>

pub fn try_new( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Box<dyn Array>, validity: Option<Bitmap> ) -> PolarsResult<Self>

Creates a new ListArray.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either crate::datatypes::PhysicalType::List or crate::datatypes::PhysicalType::LargeList.

The data_type’s inner field’s data type is not equal to values.data_type.

Implementation

This function is O(1)

pub fn new( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Box<dyn Array>, validity: Option<Bitmap> ) -> Self

Creates a new ListArray.

Panics

This function panics iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either crate::datatypes::PhysicalType::List or crate::datatypes::PhysicalType::LargeList.

The data_type’s inner field’s data type is not equal to values.data_type.

Implementation

This function is O(1)

pub fn new_empty(data_type: ArrowDataType) -> Self

Returns a new empty ListArray.

pub fn new_null(data_type: ArrowDataType, length: usize) -> Self

Returns a new null ListArray.

impl<O: Offset> ListArray<O>

pub fn slice(&mut self, offset: usize, length: usize)

Slices this ListArray.

Panics

panics iff offset + length >= self.len()

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this ListArray.

Safety

The caller must ensure that offset + length < self.len().

pub fn sliced(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn sliced_unchecked(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> Self

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

impl<O: Offset> ListArray<O>

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> Box<dyn Array>

Returns the element at index i

Panic

Panics iff i >= self.len()

pub unsafe fn value_unchecked(&self, i: usize) -> Box<dyn Array>

Returns the element at index i as &str

Safety

Assumes that the i < self.len.

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn offsets(&self) -> &OffsetsBuffer<O>

The offsets [Buffer].

pub fn values(&self) -> &Box<dyn Array>

The values.

impl<O: Offset> ListArray<O>

pub fn default_datatype(data_type: ArrowDataType) -> ArrowDataType

Returns a default ArrowDataType: inner field is named “item” and is nullable

pub fn get_child_field(data_type: &ArrowDataType) -> &Field

Returns a the inner Field

Panics

Panics iff the logical type is not consistent with this struct.

pub fn try_get_child(data_type: &ArrowDataType) -> PolarsResult<&Field>

Returns a the inner Field

Errors

Panics iff the logical type is not consistent with this struct.

pub fn get_child_type(data_type: &ArrowDataType) -> &ArrowDataType

Returns a the inner ArrowDataType

Panics

Panics iff the logical type is not consistent with this struct.