Type Alias polars_arrow::legacy::prelude::LargeStringArray

pub type LargeStringArray = Utf8Array<i64>;

Aliased Type

struct LargeStringArray { /* private fields */ }

Implementations

impl<O: Offset> Utf8Array<O>

pub fn try_new( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> PolarsResult<Self>

Returns a Utf8Array created from its internal representation.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(N) - checking utf8 is O(N)

pub fn from_slice<T: AsRef<str>, P: AsRef<[T]>>(slice: P) -> Self

Returns a Utf8Array from a slice of &str.

A convenience method that uses Self::from_trusted_len_values_iter.

pub fn from<T: AsRef<str>, P: AsRef<[Option<T>]>>(slice: P) -> Self

Returns a new Utf8Array from a slice of &str.

A convenience method that uses Self::from_trusted_len_iter.

pub fn iter(&self) -> ZipValidity<&str, Utf8ValuesIter<'_, O>, BitmapIter<'_>>

Returns an iterator of Option<&str>

pub fn values_iter(&self) -> Utf8ValuesIter<'_, O>

Returns an iterator of &str

pub fn len(&self) -> usize

Returns the length of this array

pub fn value(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Panic

This function panics iff i >= self.len.

pub unsafe fn value_unchecked(&self, i: usize) -> &str

Returns the value of the element at index i, ignoring the array’s validity.

Safety

This function is safe iff i < self.len.

pub fn get(&self, i: usize) -> Option<&str>

Returns the element at index i or None if it is null

Panics

iff i >= self.len()

pub fn data_type(&self) -> &ArrowDataType

Returns the ArrowDataType of this array.

pub fn values(&self) -> &Buffer<u8>

Returns the values of this Utf8Array.

pub fn offsets(&self) -> &OffsetsBuffer<O>

Returns the offsets of this Utf8Array.

pub fn validity(&self) -> Option<&Bitmap>

The optional validity.

pub fn slice(&mut self, offset: usize, length: usize)

Slices this Utf8Array.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn slice_unchecked(&mut self, offset: usize, length: usize)

Slices this Utf8Array.

Implementation

This function is O(1)

Safety

The caller must ensure that offset + length <= self.len().

pub fn sliced(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Panics

iff offset + length > self.len().

pub unsafe fn sliced_unchecked(self, offset: usize, length: usize) -> Self

Returns this array sliced.

Implementation

This function is O(1).

Safety

The caller must ensure that offset + length <= self.len().

pub fn with_validity(self, validity: Option<Bitmap>) -> Self

Returns this array with a new validity.

Panic

Panics iff validity.len() != self.len().

pub fn set_validity(&mut self, validity: Option<Bitmap>)

Sets the validity of this array.

Panics

This function panics iff values.len() != self.len().

pub fn boxed(self) -> Box<dyn Array>

Boxes this array into a Box<dyn Array>.

pub fn arced(self) -> Arc<dyn Array>

Arcs this array into a std::sync::Arc<dyn Array>.

pub fn into_inner( self ) -> (ArrowDataType, OffsetsBuffer<O>, Buffer<u8>, Option<Bitmap>)

Returns its internal representation

pub fn into_mut(self) -> Either<Self, MutableUtf8Array<O>>

Try to convert this Utf8Array to a MutableUtf8Array

pub fn new_empty(data_type: ArrowDataType) -> Self

Returns a new empty Utf8Array.

The array is guaranteed to have no elements nor validity.

pub fn new_null(data_type: ArrowDataType, length: usize) -> Self

Returns a new Utf8Array whose all slots are null / None.

pub fn default_data_type() -> ArrowDataType

Returns a default ArrowDataType of this array, which depends on the generic parameter O: DataType::Utf8 or DataType::LargeUtf8

pub unsafe fn try_new_unchecked( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> PolarsResult<Self>

Creates a new Utf8Array without checking for offsets monotinicity nor utf8-validity

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

Safety

This function is unsound iff:

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(1)

pub fn new( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> Self

Creates a new Utf8Array.

Panics

This function panics iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(N) - checking utf8 is O(N)

pub unsafe fn new_unchecked( data_type: ArrowDataType, offsets: OffsetsBuffer<O>, values: Buffer<u8>, validity: Option<Bitmap> ) -> Self

Creates a new Utf8Array without checking for offsets monotinicity.

Errors

This function returns an error iff:

The last offset is not equal to the values’ length.

the validity’s length is not equal to offsets.len().

The data_type’s crate::datatypes::PhysicalType is not equal to either Utf8 or LargeUtf8.

Safety

This function is unsound iff:

the offsets are not monotonically increasing

The values between two consecutive offsets are not valid utf8

Implementation

This function is O(1)

pub fn from_trusted_len_values_iter<T: AsRef<str>, I: TrustedLen<Item = T>>( iterator: I ) -> Self

Returns a (non-null) Utf8Array created from a TrustedLen of &str.

Implementation

This function is O(N)

pub fn from_iter_values<T: AsRef<str>, I: Iterator<Item = T>>( iterator: I ) -> Self

Creates a new Utf8Array from a Iterator of &str.

pub unsafe fn from_trusted_len_iter_unchecked<I, P>(iterator: I) -> Selfwhere P: AsRef<str>, I: Iterator<Item = Option<P>>,

Creates a Utf8Array from an iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn from_trusted_len_iter<I, P>(iterator: I) -> Selfwhere P: AsRef<str>, I: TrustedLen<Item = Option<P>>,

Creates a Utf8Array from an iterator of trusted length.

pub unsafe fn try_from_trusted_len_iter_unchecked<E, I, P>( iterator: I ) -> Result<Self, E>where P: AsRef<str>, I: IntoIterator<Item = Result<Option<P>, E>>,

Creates a Utf8Array from an falible iterator of trusted length.

Safety

The iterator must be TrustedLen. I.e. that size_hint().1 correctly reports its length.

pub fn try_from_trusted_len_iter<E, I, P>(iter: I) -> Result<Self, E>where P: AsRef<str>, I: TrustedLen<Item = Result<Option<P>, E>>,

Creates a Utf8Array from an fallible iterator of trusted length.

pub fn apply_validity<F: FnOnce(Bitmap) -> Bitmap>(&mut self, f: F)

Applies a function f to the validity of this array.

This is an API to leverage clone-on-write

Panics

This function panics if the function f modifies the length of the Bitmap.