polars_arrow

Module record_batch

Contains RecordBatchT, a container of Array where every array has the same length.

Structs

RecordBatchT

A vector of trait objects of Array where every item has the same length, RecordBatchT::len.

Type Aliases

RecordBatch