polars_arrow::types

Trait AlignedBytesCast

pub unsafe trait AlignedBytesCast<B: AlignedBytes>: Pod { }

Define that a type has the same byte alignment and size as B.

Safety

This is safe to implement if both types have the same alignment and size.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.