Trait polars_arrow::util::total_ord::TotalEq

pub trait TotalEq {
    // Required method
    fn tot_eq(&self, other: &Self) -> bool;

    // Provided method
    fn tot_ne(&self, other: &Self) -> bool { ... }
}

Alternative trait for Eq. By consistently using this we can still be generic w.r.t Eq while getting a total ordering for floats.

Required Methods

fn tot_eq(&self, other: &Self) -> bool

Provided Methods

fn tot_ne(&self, other: &Self) -> bool