Trait polars_arrow::util::total_ord::TotalOrd

pub trait TotalOrd: TotalEq {
    // Required method
    fn tot_cmp(&self, other: &Self) -> Ordering;

    // Provided methods
    fn tot_lt(&self, other: &Self) -> bool { ... }
    fn tot_gt(&self, other: &Self) -> bool { ... }
    fn tot_le(&self, other: &Self) -> bool { ... }
    fn tot_ge(&self, other: &Self) -> bool { ... }
}

Alternative trait for Ord. By consistently using this we can still be generic w.r.t Ord while getting a total ordering for floats.

Required Methods

fn tot_cmp(&self, other: &Self) -> Ordering

Provided Methods

fn tot_lt(&self, other: &Self) -> bool

fn tot_gt(&self, other: &Self) -> bool

fn tot_le(&self, other: &Self) -> bool

fn tot_ge(&self, other: &Self) -> bool