polars_arrow::array::growable

Trait Growable

pub trait Growable<'a> {
    // Required methods
    unsafe fn extend(&mut self, index: usize, start: usize, len: usize);
    fn extend_validity(&mut self, additional: usize);
    fn len(&self) -> usize;
    fn as_box(&mut self) -> Box<dyn Array>;

    // Provided methods
    unsafe fn extend_copies(
        &mut self,
        index: usize,
        start: usize,
        len: usize,
        copies: usize,
    ) { ... }
    fn as_arc(&mut self) -> Arc<dyn Array> { ... }
}

Describes a struct that can be extended from slices of other pre-existing Arrays. This is used in operations where a new array is built out of other arrays, such as filter and concatenation.

Required Methods

unsafe fn extend(&mut self, index: usize, start: usize, len: usize)

Extends this Growable with elements from the bounded Array at index index from a slice starting at start and length len.

Safety

Doesn’t do any bound checks.

fn extend_validity(&mut self, additional: usize)

Extends this Growable with null elements, disregarding the bound arrays

Safety

Doesn’t do any bound checks

fn len(&self) -> usize

The current length of the Growable.

fn as_box(&mut self) -> Box<dyn Array>

Converts this Growable to an Box<dyn Array>, thereby finishing the mutation. Self will be empty after such operation

Provided Methods

unsafe fn extend_copies( &mut self, index: usize, start: usize, len: usize, copies: usize, )

Same as extend, except it repeats the extension copies times.

Safety

Doesn’t do any bound checks.

fn as_arc(&mut self) -> Arc<dyn Array>

Converts this Growable to an Arc<dyn Array>, thereby finishing the mutation. Self will be empty after such operation.