Trait polars_arrow::array::ListFromIter

pub trait ListFromIter {
    unsafe fn from_iter_primitive_trusted_len<T, P, I>(
        iter: I
    ) -> LargeListArray
    where
        T: ArrowNumericType,
        P: IntoIterator<Item = Option<T::Native>>,
        I: IntoIterator<Item = Option<P>>,
    { ... }

    unsafe fn from_iter_bool_trusted_len<I, P>(iter: I) -> LargeListArray
    where
        I: IntoIterator<Item = Option<P>>,
        P: IntoIterator<Item = Option<bool>>,
    { ... }

    unsafe fn from_iter_utf8_trusted_len<I, P, Ref>(
        iter: I, 
        n_elements: usize
    ) -> LargeListArray
    where
        I: IntoIterator<Item = Option<P>>,
        P: IntoIterator<Item = Option<Ref>>,
        Ref: AsRef<str>,
    { ... }
}

Provided methods

unsafe fn from_iter_primitive_trusted_len<T, P, I>(iter: I) -> LargeListArray where
T: ArrowNumericType,
P: IntoIterator<Item = Option<T::Native>>,
I: IntoIterator<Item = Option<P>>,

Create a list-array from an iterator. Used in groupby agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_bool_trusted_len<I, P>(iter: I) -> LargeListArray where
I: IntoIterator<Item = Option<P>>,
P: IntoIterator<Item = Option<bool>>,

Create a list-array from an iterator. Used in groupby agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

unsafe fn from_iter_utf8_trusted_len<I, P, Ref>(
iter: I,
n_elements: usize
) -> LargeListArray where
I: IntoIterator<Item = Option<P>>,
P: IntoIterator<Item = Option<Ref>>,
Ref: AsRef<str>,

Create a list-array from an iterator. Used in groupby agg-list

Safety

Will produce incorrect arrays if size hint is incorrect.

Loading content...