Trait polars_arrow::array::TryExtendFromSelf

pub trait TryExtendFromSelf {
    // Required method
    fn try_extend_from_self(&mut self, other: &Self) -> PolarsResult<()>;
}

A trait describing the ability of a struct to extend from a reference of itself. Specialization of TryExtend.

Required Methods

fn try_extend_from_self(&mut self, other: &Self) -> PolarsResult<()>

Tries to extend itself with elements from other, failing only on overflow.