polars_arrow::compute

Module boolean

null-preserving operators such as and, or and not.

Functions

all

Returns whether all values in the array are true.

and

Performs && operation on two BooleanArray, combining the validities.

and_scalar

Performs AND operation on an array and a scalar value. If either left or right value is null then the result is also null.

any

Returns whether any of the values in the array are true.

is_not_null

Returns a non-null BooleanArray with whether each value of the array is not null.

is_null

Returns a non-null BooleanArray with whether each value of the array is null.

not

Performs unary NOT operation on an arrays. If value is null then the result is also null.

or

Performs || operation on two BooleanArray, combining the validities.

or_scalar

Performs OR operation on an array and a scalar value. If either left or right value is null then the result is also null.