Module polars_arrow::compute::take

Functions

take_binary_iter_unchecked⚠

Safety

take_binary_opt_iter_unchecked⚠

Safety

take_binary_unchecked⚠

Safety

take_bool_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.

take_bool_opt_iter_unchecked⚠

Take kernel for single chunk and an iterator as index.

take_no_null_binary_iter_unchecked⚠

Safety

take_no_null_binary_opt_iter_unchecked⚠

Safety

take_no_null_bool_iter_unchecked⚠

Take kernel for single chunk without nulls and an iterator as index.

take_no_null_bool_opt_iter_unchecked⚠

Take kernel for single chunk without null values and an iterator as index that may produce None values.

take_no_null_primitive_iter_unchecked⚠

Take kernel for single chunk without nulls and an iterator as index.

take_no_null_primitive_opt_iter_unchecked⚠

Take kernel for a single chunk without nulls and an iterator that can produce None values. This is used in join operations.

take_no_null_primitive_unchecked⚠

Take kernel for single chunk without nulls and arrow array as index.

take_no_null_utf8_iter_unchecked⚠

Safety

take_no_null_utf8_opt_iter_unchecked⚠

Safety

take_primitive_iter_unchecked⚠

Take kernel for a single chunk with null values and an iterator as index.

take_primitive_opt_iter_unchecked⚠

Take kernel for a single chunk and an iterator that can produce None values. This is used in join operations.

take_primitive_unchecked⚠

Take kernel for single chunk with nulls and arrow array as index that may have nulls.

take_unchecked⚠

Safety

take_utf8_iter_unchecked⚠

Safety

take_utf8_opt_iter_unchecked⚠

Safety

take_utf8_unchecked⚠

Safety

take_value_indices_from_list⚠

Forked and adapted from arrow-rs This is faster because it does no bounds checks and allocates directly into aligned memory