Module polars_arrow::compute::temporal

Defines temporal kernels for time and date related functions.

Functions

can_day

Checks if an array of type datatype can perform day operation

can_hour

Checks if an array of type datatype can perform hour operation

can_iso_week

Checks if an array of type data_type can perform ISO week operation

can_minute

Checks if an array of type datatype can perform minute operation

can_month

Checks if an array of type datatype can perform month operation

can_nanosecond

Checks if an array of type datatype can perform nanosecond operation

can_second

Checks if an array of type datatype can perform second operation

can_weekday

Checks if an array of type data_type can perform weekday operation

can_year

Checks if an array of type datatype can perform year operation

day

Extracts the days of a temporal array as PrimitiveArray<u32>. Value ranges from 1 to 32 (Last day depends on month). Use can_day to check if this operation is supported for the target ArrowDataType.

hour

Extracts the hours of a temporal array as PrimitiveArray<u32>. Value ranges from 0 to 23. Use can_hour to check if this operation is supported for the target ArrowDataType.

iso_week

Extracts ISO week of a temporal array as PrimitiveArray<u32> Value ranges from 1 to 53 (Last week depends on the year). Use can_iso_week to check if this operation is supported for the target ArrowDataType

minute

Extracts the minutes of a temporal array as PrimitiveArray<u32>. Value ranges from 0 to 59. Use can_minute to check if this operation is supported for the target ArrowDataType.

month

Extracts the months of a temporal array as PrimitiveArray<u32>. Value ranges from 1 to 12. Use can_month to check if this operation is supported for the target ArrowDataType.

nanosecond

Extracts the nanoseconds of a temporal array as PrimitiveArray<u32>. Use can_nanosecond to check if this operation is supported for the target ArrowDataType.

second

Extracts the seconds of a temporal array as PrimitiveArray<u32>. Value ranges from 0 to 59. Use can_second to check if this operation is supported for the target ArrowDataType.

weekday

Extracts weekday of a temporal array as PrimitiveArray<u32>. Monday is 1, Tuesday is 2, …, Sunday is 7. Use can_weekday to check if this operation is supported for the target ArrowDataType

year

Extracts the years of a temporal array as PrimitiveArray<i32>. Use can_year to check if this operation is supported for the target ArrowDataType.