polars_arrow::legacy::kernels::rolling::nulls

Trait RollingAggWindowNulls

pub trait RollingAggWindowNulls<'a, T: NativeType> {
    // Required methods
    unsafe fn new(
        slice: &'a [T],
        validity: &'a Bitmap,
        start: usize,
        end: usize,
        params: Option<RollingFnParams>,
    ) -> Self;
    unsafe fn update(&mut self, start: usize, end: usize) -> Option<T>;
    fn is_valid(&self, min_periods: usize) -> bool;
}

Required Methods

unsafe fn new( slice: &'a [T], validity: &'a Bitmap, start: usize, end: usize, params: Option<RollingFnParams>, ) -> Self

Safety

start and end must be in bounds for slice and validity

unsafe fn update(&mut self, start: usize, end: usize) -> Option<T>

Safety

start and end must be in bounds of slice and bitmap

fn is_valid(&self, min_periods: usize) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.