polars_arrow::legacy::kernels::set

Function scatter_single_non_null

pub fn scatter_single_non_null<T, I>(
    array: &PrimitiveArray<T>,
    idx: I,
    set_value: T,
    dtype: ArrowDataType,
) -> PolarsResult<PrimitiveArray<T>>
where
    T: NativeType,
    I: IntoIterator<Item = IdxSize>,

Efficiently sets value at the indices from the iterator to set_value. The new array is initialized with a memcpy from the old values.