Trait polars_arrow::trusted_len::PushUnchecked

pub trait PushUnchecked<T> {
    // Required methods
    unsafe fn push_unchecked(&mut self, value: T);
    unsafe fn push_unchecked_no_len_set(&mut self, value: T);
    unsafe fn extend_trusted_len_unchecked<I: IntoIterator<Item = T>>(
        &mut self,
        iter: I
    );
    unsafe fn from_trusted_len_iter_unchecked<I: IntoIterator<Item = T>>(
        iter: I
    ) -> Self;

    // Provided methods
    fn extend_trusted_len<I: IntoIterator<Item = T, IntoIter = J>, J: TrustedLen>(
        &mut self,
        iter: I
    ) { ... }
    fn from_trusted_len_iter<I: IntoIterator<Item = T, IntoIter = J>, J: TrustedLen>(
        iter: I
    ) -> Self
       where Self: Sized { ... }
}

Required Methods

unsafe fn push_unchecked(&mut self, value: T)

Will push an item and not check if there is enough capacity

Safety

Caller must ensure the array has enough capacity to hold T.

unsafe fn push_unchecked_no_len_set(&mut self, value: T)

Will push an item and not check if there is enough capacity nor update the array’s length

Safety

Caller must ensure the array has enough capacity to hold T. Caller must update the length when its done updating the vector.

unsafe fn extend_trusted_len_unchecked<I: IntoIterator<Item = T>>( &mut self, iter: I )

Safety

Caller must ensure the iterators reported length is correct

unsafe fn from_trusted_len_iter_unchecked<I: IntoIterator<Item = T>>( iter: I ) -> Self

Safety

Caller must ensure the iterators reported length is correct

Provided Methods

fn extend_trusted_len<I: IntoIterator<Item = T, IntoIter = J>, J: TrustedLen>( &mut self, iter: I )

Extend the array with an iterator who’s length can be trusted

fn from_trusted_len_iter<I: IntoIterator<Item = T, IntoIter = J>, J: TrustedLen>( iter: I ) -> Selfwhere Self: Sized,