Trait polars_arrow::trusted_len::PushUnchecked

pub trait PushUnchecked<T> {
    unsafe fn push_unchecked(&mut self, value: T);

    unsafe fn push_unchecked_no_len_set(&mut self, value: T);

    fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(
        &mut self, 
        iter: I
    );

    fn from_trusted_len_iter<I: IntoIterator<Item = T> + TrustedLen>(
        iter: I
    ) -> Self;
}

Required methods

unsafe fn push_unchecked(&mut self, value: T)

Will push an item and not check if there is enough capacity

Safety

Caller must ensure the array has enough capacity to hold T.

unsafe fn push_unchecked_no_len_set(&mut self, value: T)

Will push an item and not check if there is enough capacity nor update the array’s length

Safety

Caller must ensure the array has enough capacity to hold T. Caller must update the length when its done updating the vector.

fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(
&mut self,
iter: I
)

Extend the array with an iterator who’s length can be trusted

fn from_trusted_len_iter<I: IntoIterator<Item = T> + TrustedLen>(
iter: I
) -> Self