Struct polars_arrow::vec::AlignedVec

pub struct AlignedVec<T: ArrowNativeType> {
    pub inner: Vec<T>,
    // some fields omitted
}

A Vec wrapper with a memory alignment equal to Arrow’s primitive arrays. Can be useful in creating a new ChunkedArray or Arrow Primitive array without copying.

Fields
inner: Vec<T>
Implementations

impl<T: Copy + ArrowNativeType> AlignedVec<T>

pub fn new_from_slice(other: &[T]) -> Self

Uses a memcpy to initialize this AlignedVec

impl<T: Clone + ArrowNativeType> AlignedVec<T>

pub fn resize(&mut self, new_len: usize, value: T)

pub fn extend_from_slice(&mut self, other: &[T])

impl<T: ArrowNativeType> AlignedVec<T>

pub fn with_capacity(size: usize) -> Self

Create a new Vec where first bytes memory address has an alignment of 64 bytes, as described by arrow spec. Read more: https://github.com/rust-ndarray/ndarray/issues/771

pub fn is_empty(&self) -> bool

pub fn reserve(&mut self, additional: usize)

pub fn len(&self) -> usize

pub unsafe fn from_ptr(ptr: usize, len: usize, capacity: usize) -> Self

Create a new aligned vec from a ptr.

Safety

The ptr should be 64 byte aligned and len and capacity should be correct otherwise it is UB.

pub fn push(&mut self, value: T)

Push at the end of the Vec. This is unsafe because a push when the capacity of the inner Vec is reached will reallocate the Vec without the alignment, leaving this destructor’s alignment incorrect

pub unsafe fn set_len(&mut self, new_len: usize)

Set the length of the underlying Vec.

Safety

new_len must be less than or equal to capacity.

The elements at old_len..new_len must be initialized.

pub fn as_ptr(&self) -> *const T

pub fn as_mut_ptr(&mut self) -> *mut T

pub fn as_mut_slice(&mut self) -> &mut [T]

pub fn as_slice(&self) -> &[T]

pub fn capacity(&self) -> usize

pub fn shrink_to_fit(&mut self)

pub fn into_arrow_buffer(self) -> Buffer

Transform this array to an Arrow Buffer.

pub fn into_primitive_array<A: ArrowPrimitiveType>(
self,
null_buf: Option<Buffer>
) -> PrimitiveArray<A>

pub fn extend<I: IntoIterator<Item = T>>(&mut self, iter: I)

Panic

Must be a trusted len iterator or else it will panic

pub unsafe fn extend_trusted_len_unchecked<I: IntoIterator<Item = T>>(
&mut self,
iter: I
)

Extend this Vector with an iterator whose length can be trusted.

Safety

iterator must be TrustedLen

pub fn extend_trusted_len<I: IntoIterator<Item = T> + TrustedLen>(
&mut self,
iter: I
)